Refactor the handle of --no-auto-commit to support --originalid
When importing with --originalid, we need to add ticket custom field
values using RT API, which calls $RT::Handle->BeginTransaction in
RT::CustomField::AddValueForObject. But if AutoCommit if off, the
BeginTransaction would error out at least with MySQL:

    DBD::mysql::db begin_work failed: Already in a transaction

This commit refactors code to not turn off AutoCommit globally, but adds
an additional transaction for each data file, which implicitly turns off
AutoCommit and is also compatible with other BeginTransaction calls.