Exempt some format strings from HTML::Gumbo
HTML::Gumbo undoes some of the escaping applied by EscapeHTML that is
necessary for properly displaying apostrophes in column titles.

This adds a flag to the scrubbing methods to skip the structure check in
cases, such as format strings, where we do not ever expect to have that
type of issue, and where side-effects of doing that check cause
problems.